+++ src-opam/windows.ml
@@ -120,14 +120,15 @@ let ocaml_for_windows_package_exn ~switch ~port ~arch =
   let _, pkgver = Ocaml_version.Opam.V2.package switch in
   ("ocaml-variants", pkgver ^ "+" ^ variant)
 
-let git_init ~name ~email ~opam_repository =
+let git_init ~name ~email ~repos =
   String.concat " && "
-    [
-      sprintf "git config --global user.email '%s'" email;
-      sprintf "git config --global user.name '%s'" name;
-      "git config --system core.longpaths true";
-      sprintf "git config --global --add safe.directory %s" opam_repository;
-    ]
+    (sprintf "git config --global user.email '%s'" email
+    :: sprintf "git config --global user.name '%s'" name
+    :: "git config --system core.longpaths true"
+    :: List.map
+         (fun repo ->
+           sprintf "git config --global --add safe.directory %s" repo)
+         repos)
 
 module Cygwin = struct
   type cyg = { root : string; site : string; args : string list }
@@ -212,10 +213,11 @@ module Cygwin = struct
   let update ?(cyg = default) () = cygsetup ~cyg ~upgrade:true "" |> cleanup
 
   let setup_env ~cyg =
-    env [ ("CYGWIN", "winsymlinks:native") ]
+    env [ ("CYGWIN", "nodosfilewarning winsymlinks:native") ]
     @@ prepend_path (List.map (( ^ ) cyg.root) [ {|\bin|} ])
 
-  let install_from_release ?(cyg = default) ?(extra = []) () =
+  let install_from_release ?(cyg = default) ?(msvs_tools = false) ?(extra = [])
+      () =
     setup_env ~cyg
     @@ add
          ~src:[ "https://www.cygwin.com/setup-x86_64.exe" ]
@@ -223,11 +225,10 @@ module Cygwin = struct
          ()
     @@ install_cygsympathy_from_source cyg
     @@ install ~cyg extra
-    @@ install_msvs_tools_from_source cyg
+    @@ (if msvs_tools then install_msvs_tools_from_source cyg else empty)
     @@ run
          {|awk -i inplace "/(^#)|(^$)/{print;next}{$4=""noacl,""$4; print}" %s\etc\fstab|}
          cyg.root
-    @@ remove_system_attribute (cyg.root ^ {|\dev|})
 
   let setup ?(cyg = default) ?from () =
     (match from with
@@ -237,17 +238,22 @@ module Cygwin = struct
     | None -> empty)
     @@ workdir {|%s\home\opam|} cyg.root
 
-  let cygwin_packages ?(extra = []) ?(flexdll_version = "0.39-1") () =
+  let cygwin_packages ?(flexdll_version = "0.39-1") () =
     (* 2021-03-19: flexdll 0.39 is required, but is in Cygwin testing *)
-    "make" :: "diffutils" :: "ocaml" :: "gcc-core" :: "git" :: "patch" :: "m4"
-    :: "cygport"
-    :: ("flexdll=" ^ flexdll_version)
-    :: extra
+    [
+      "make";
+      "diffutils";
+      "ocaml";
+      "gcc-core";
+      "git";
+      "patch";
+      "m4";
+      "cygport";
+      "flexdll=" ^ flexdll_version;
+    ]
 
-  let mingw_packages ?(extra = []) () =
-    "make" :: "diffutils" :: "mingw64-x86_64-gcc-core" :: extra
-
-  let msvc_packages ?(extra = []) () = "make" :: "diffutils" :: extra
+  let mingw_packages = [ "make"; "diffutils"; "mingw64-x86_64-gcc-core" ]
+  let msvc_packages = [ "make"; "diffutils" ]
 
   let ocaml_for_windows_packages ?cyg ?(extra = []) ?(version = "0.0.0.2") () =
     let packages =
@@ -269,10 +275,9 @@ module Cygwin = struct
 
   module Git = struct
     let init ?(cyg = default) ?(name = "Docker") ?(email = "docker@example.com")
-        () =
+        ?(repos = [ "/home/opam/opam-repository" ]) () =
       env [ ("HOME", cyg.root ^ {|\home\opam|}) ]
-      @@ run_sh ~cyg "%s"
-           (git_init ~email ~name ~opam_repository:"/home/opam/opam-repository")
+      @@ run_sh ~cyg "%s" (git_init ~email ~name ~repos)
   end
 end
 
@@ -344,9 +349,8 @@ module Winget = struct
     | _ -> install [ "Git.Git" ] @@ maybe install extra
 
   module Git = struct
-    let init ?(name = "Docker") ?(email = "docker@example.com") () =
-      run "%s"
-        (git_init ~email ~name
-           ~opam_repository:"C:/cygwin64/home/opam/opam-repository")
+    let init ?(name = "Docker") ?(email = "docker@example.com")
+        ?(repos = [ "C:/cygwin64/home/opam/opam-repository" ]) () =
+      run "%s" (git_init ~email ~name ~repos)
   end
 end